cloud functions/fetch_tweets/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.16.0
google-cloud-language==2.9.1
google-cloud-secret-manager==2.16.1
google-cloud-bigquery==3.10.0
requests==2.28.2